Enable version validation in schema9 blobs.
